+++ Demo/API/Assets/WX-WASM-SDK/CHANGELOG.md
@@ -6,6 +6,31 @@ Removed - 删除功能/接口
 Fixed - 修复问题
 Others - 其他 
 -->
+2022-12-7
+### Feature
+* 支持指定Node路径，MiniGameConfig.asset-CompileOption-CustomNodePath指定自定义node安装目录
+* 编译选项增加CleanBuild(仅支持2021以上版本)
+* 适配插件版本升级到1.1.3
+### Fixed
+* 修复UnityAudio适配InnerAudio时，静音状态对长音频在切换场景后无效的问题
+
+2022-11-30
+### Feature
+* 微信压缩纹理支持WebGL2.0
+* 微信压缩纹理对.svn .git目录忽略 
+
+2022-11-23
+### Feature
+*  导出增加sbrk函数，2021无需profiling-memory查看DynamicMemory
+
+### Fixed
+*  UDPClient使用connect+write, 修复send接口性能问题
+
+ 2022-11-17
+### Feature
+*  微信压缩纹理优化，支持ASTC使用非4倍数尺寸
+*  新增录屏接口功能
+
 2022-11-9
 ### Feature
 *  Unity Audio压缩纹理，支持小游戏InnerAudio